refactor: packages ..model to ..model/events
